Make quiche work at 32bit OS
For stream handling, "offset" (or similar data such as max_data)
is now defined as u64 for consistency between different
architecture, mainly for 32bit arch. Also keeping "len" as usize.

When type conversion is necessary, mostly it's usize -> u64
for not losing existing bits.

+++ src/frame.rs
@@ -33,8 +33,8 @@ use crate::packet;
 use crate::ranges;
 use crate::stream;
 
-pub const MAX_CRYPTO_OVERHEAD: usize = 8;
-pub const MAX_STREAM_OVERHEAD: usize = 12;
+pub const MAX_CRYPTO_OVERHEAD: u64 = 8;
+pub const MAX_STREAM_OVERHEAD: u64 = 12;
 pub const MAX_STREAM_SIZE: u64 = 1 << 62;
 
 #[derive(Clone, PartialEq)]
@@ -176,7 +176,7 @@ impl Frame {
             },
 
             0x06 => {
-                let offset = b.get_varint()? as usize;
+                let offset = b.get_varint()?;
                 let data = b.get_bytes_with_varint_length()?;
                 let data = stream::RangeBuf::from(data.as_ref(), offset, false);
 
@@ -850,7 +850,7 @@ fn parse_stream_frame(ty: u64, b: &mut octets::Octets) -> Result<Frame> {
     let stream_id = b.get_varint()?;
 
     let offset = if first & 0x04 != 0 {
-        b.get_varint()? as usize
+        b.get_varint()?
     } else {
         0
     };
@@ -861,7 +861,7 @@ fn parse_stream_frame(ty: u64, b: &mut octets::Octets) -> Result<Frame> {
         b.cap()
     };
 
-    if offset + len >= MAX_STREAM_SIZE  {
+    if offset + len as u64 >= MAX_STREAM_SIZE {
         return Err(Error::InvalidFrame);
     }
